+++ src/main/java/blackjack/model/cards/BestScoreCards.java
@@ -2,7 +2,6 @@ package blackjack.model.cards;
 
 import blackjack.model.card.Card;
 import blackjack.model.card.Rank;
-import blackjack.model.blackjack.Score;
 
 final class BestScoreCards extends ScoreCards {
