+++ src/main/java/blackjack/model/cards/Score.java
@@ -1,4 +1,4 @@
-package blackjack.model.blackjack;
+package blackjack.model.cards;
 
 import java.util.Objects;
