+++ src/main/java/blackjack/model/player/Dealer.java
@@ -2,7 +2,7 @@ package blackjack.model.player;
 
 import blackjack.model.card.Card;
 import blackjack.model.blackjack.Result;
-import blackjack.model.blackjack.Score;
+import blackjack.model.cards.Score;
 import blackjack.model.cards.Cards;
 import blackjack.model.cards.ScoreCards;
