+++ src/main/java/blackjack/model/player/Player.java
@@ -1,7 +1,7 @@
 package blackjack.model.player;
 
 import blackjack.model.card.Card;
-import blackjack.model.blackjack.Score;
+import blackjack.model.cards.Score;
 import blackjack.model.cards.Cards;
 import blackjack.model.cards.ScoreCards;
