+++ src/main/java/blackjack/view/OutputView.java
@@ -2,7 +2,7 @@ package blackjack.view;
 
 import blackjack.model.blackjack.Record;
 import blackjack.model.blackjack.Result;
-import blackjack.model.blackjack.Score;
+import blackjack.model.cards.Score;
 import blackjack.model.card.Card;
 import blackjack.model.cards.Cards;
 import blackjack.model.player.Name;
