+++ src/test/java/blackjack/model/blackjack/ScoreTest.java
@@ -2,6 +2,7 @@ package blackjack.model.blackjack;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.model.cards.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
