+++ src/test/java/blackjack/model/player/DealerTest.java
@@ -8,7 +8,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.model.blackjack.Result;
-import blackjack.model.blackjack.Score;
+import blackjack.model.cards.Score;
 import blackjack.model.card.Card;
 import blackjack.model.card.Rank;
 import java.util.stream.Stream;
