+++ src/test/java/blackjack/model/player/PlayerTest.java
@@ -6,7 +6,7 @@ import static blackjack.model.card.Suit.SPADE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.model.blackjack.Score;
+import blackjack.model.cards.Score;
 import blackjack.model.card.Card;
 import blackjack.model.card.Rank;
 import java.util.stream.Stream;
